feat: Add `dump` command
Rename CLI, previous single command is now `migrate`
