+++ Cargo.toml
@@ -7,7 +7,7 @@ license = "Apache-2.0"
 publish = false
 
 [[bin]]
-name = "ds-migrate"
+name = "declare-schema"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -26,4 +26,4 @@ service_conventions = { version = "0.0.16", features = ["tracing", "oidc"]}
 sqlx = { version = "0.7.4", features = ["runtime-tokio", "tls-rustls", "postgres", "macros", "uuid", "time", "chrono", "rust_decimal"] }
 uuid = { version = "1.8.0", features = ["v4", "serde"] }
 thiserror = "1.0.59"
-sqlparser = { version="0.46.0", git = "https://github.com/philipcristiano/sqlparser-rs.git", branch = "create-table-struct"}
+sqlparser = {  git = "https://github.com/sqlparser-rs/sqlparser-rs.git", branch = "main"}
