+++ Dockerfile
@@ -6,7 +6,7 @@ RUN cargo install --path .
 
 FROM debian:bookworm-slim
 RUN apt-get update && apt-get install -y procps ca-certificates && rm -rf /var/lib/apt/lists/*
-COPY --from=builder /usr/local/cargo/bin/ds-migrate /usr/local/bin/ds-migrate
+COPY --from=builder /usr/local/cargo/bin/declare-schema /usr/local/bin/declare-schema
 
 
-ENTRYPOINT ["/usr/local/bin/ds-migrate"]
+ENTRYPOINT ["/usr/local/bin/declare-schema"]
